// src/app/shared/interfaces/permission.ts
export class Permission{
      add_product : boolean;
      edit_product: boolean;
      delete_product: boolean;
      view_order: boolean;
      update_order: boolean;
      delete_order: boolean;
      add_user: boolean;
      update_user: boolean;
      delete_user: boolean;
      view_customer: boolean;
      add_coupon: boolean;
      delete_coupon: boolean;
      update_setting: boolean;
      add_shop: boolean;
      update_shop: boolean;
      delete_shop: boolean;
      view_shop: boolean;
}